Wizard + Share Raw CAN visibility + dns_hijack removal + WIP rollup
WIZARD (first_run_wizard.c + ui_wifi.c/h):
- Step 3 now exposes both connection paths as buttons. "Join a WiFi
  Network" applies STA mode + persists wifi-on-boot; "Start Hotspot"
  applies AP mode + persists hotspot-on-boot. Both go through a new
  public wifi_ui_show_with_preset() that applies the runtime mode AND
  saves Start-on-Boot before opening the WiFi screen.
- Step 1 gains a "Skip for good" button that marks first_run_done so
  the wizard never reappears. Card height bumped from 430 to 470 to
  give it room.

SHARE RAW CAN BUTTON (device_settings.c):
- Data Logging row was 95 px tall but the Share Raw CAN button sat at
  y=92 with height 30 — clipped entirely on device. Row bumped to 140.

DNS HIJACK REMOVAL:
- Drops dns_hijack.c/h, removes its CMakeLists entry, strips boot-time
  startup from main.c and the start/stop calls from ui_wifi.c. The
  feature wasn't actually doing what the comments claimed on phones
  and was eating heap; removing simplifies the network stack.

Also folds in WIP across can_bus_test, can_manager, ecu_presets,
ota_handler, web_server*, wifi_manager, config_store, data_logger,
widget_meter, sdkconfig, and the cloud-proxy / dev-server tooling.
This was a single uncommitted rollup so it lands as one commit.

Also adds the can_raw_logger.{c,h} + favicon.ico files that CMakeLists
has long referenced but were never tracked.

+++ main/widgets/widget_meter.c
@@ -874,7 +874,7 @@ static void _meter_from_json(widget_t *w, cJSON *in) {
 	ap = cJSON_GetObjectItemCaseSensitive(cfg, "scale_padding");
 	if (cJSON_IsNumber(ap)) md->scale_padding = (uint8_t)ap->valueint;
 	ap = cJSON_GetObjectItemCaseSensitive(cfg, "label_gap");
-	if (cJSON_IsNumber(ap)) md->label_gap = (int8_t)ap->valueint;
+	if (cJSON_IsNumber(ap)) md->label_gap = (int16_t)ap->valueint;
 	ap = cJSON_GetObjectItemCaseSensitive(cfg, "tick_label_font");
 	if (cJSON_IsString(ap) && ap->valuestring) {
 		safe_strncpy(md->tick_label_font, ap->valuestring, sizeof(md->tick_label_font));
